refactoring dirs
